Split test_collection file into smaller test files
